// apps/anuglar/src/app/pages/order/config/enums/route-names.ts
export const enum eOrderRouteNames {
  SalesOrder = 'Orders::Menu:SalesOrder',
  SalesOrderNew = 'Orders::Menu:SalesOrderNew',
  SalesOrderList = 'Orders::Menu:SalesOrderList',
  PurchaseOrder =     'Orders::Menu:PurchaseOrder',
  PurchaseOrderNew =  'Orders::Menu:PurchaseOrderNew',
  PurchaseOrderList = 'Orders::Menu:PurchaseOrderList',
}